requirements.txt
absl-py==2.2.2
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==20.1.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.2
joblib==1.5.0
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
namex==0.0.9
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.1.0
pillow==11.2.1
protobuf==4.25.7
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.12.1
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.1
requests==2.31.0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
